changed layout, added search element, fixes...
Fixed the calculated height of the canvas container.
It finally shows the whole scrollbar,
in order to achieve that we manually
run onResize() on page load.
Added the html input element for search,
it is barely configured yet though.
Lastly changed the layout to the same one used by the previous application,
to that effect the scrollbars for the window are hidden,
since both sides of the page are scrollable themselves.