Split codebase into more files...
Furthermore, I moved some event bindings from the html file to a script.
Note that this source structure is not done yet.
As I am having problems with splitting files into folders.
The mod.ts file would be the main file of the weapons folder following the rust convention.
But I need to solve that first.