Implement drawing text columns...
Previously everything was drawn manually.
This commit begins the work to make CellRenderers
which will handle most of the common use cases.
Finished the TextCellRender, which draws only text
to its cell area.
Reimplemented the affinity column using this new API.
Note that this is still not the public interface.
The user will only need to set the data and the
TreeView will take care of the rest.
Later on you will also be able to make your own
CellRenderer which will allow for drawing custom
cells like what will be needed for the sharpness column.