Removed clipping in favour of scaling down rect...
The clipping impacted the performance too much,
the search got unresponsive as a result.
This method still provides the graphical benefit
of everything fitting correctly but without the
performance downside.
Of course this means that the content of the cell can
be drawn outside of its assigned rect but this should
not be a problem for this project.