Added tree data structure for TreeView...
ATM the structure is not used yet for drawing but
it will be a better solution long term and it will allow
for proprely ordering MR kulve taroth weapons without
having to fix the data manually.
In the future we should be able to use this better
structure along with is_visible to support expanding
and collapsing tree nodes.
Also this will allow us to sort the data in tree form, by
only sorting the root nodes which should also just
work for when the data is displayed in table format
since each row is just a root node.