Consolidated tree drawing functions...
It makes more sense to do both the lines and the rows
at the same time when descending the tree.
Still need to change the underlying data and wire it up correctly.
ATM most weapons match up with exceptions to kulve.
The way to solve this is to switch to the tree data structure
and write new code for selection.