Updated details, changed data binding to model...
Previously each column of the model was just an
element of an array.
Now the model is made from an object.
The upsides of this is that each column has a nice
user given name to refer to later. Using an object
also make it easier to implement data that needs
to be stored but not displayed, like the id in this case.
We dont want to display the id but we need it for later
to query the db for the rest of the info which we
dont store in the model.