Finished prototype for sorting columns...
Right now its hardcoded for attack only.
List of things to do:
display asc/desc colored triangle in the header if not elipsized,
provide sorting for all or some select columns.
The second point might be problematic for some cols.
Because not every column has a straight forward number repr.
Name, attack, elderseal are easy.
For element the road of least resistance would be to
just compare the elemental attack value ( - paren?).
Obviously this does not allow for sorting by element.
Now that I think about it, we should be able to utilise
hidden columns to store information for comparison.
This means that sorting slots should be trivial
sicne we can just store the numeric value in hidden
and simply compare that.
Sharpness could be compared by multiplying each
sharpness tier by its damage multiplier and adding each value together.
For affinity and defense we could store the pure numeric value,
which should make for a straight forward comaprison.
For element attack we would comapre the attack
 value but sort weapons with hidden element to the bottom.