Finished sort implementation...
Abstracted the sort into a common function which
only differs in what properites it compares and what
result it should return asc or desc.
It also draws a small triangle to indicate the sort on the column.
If the header content + the triangle size is too much
for the current column width it will be expanded.
Had to modify the calculateColumn function since
clicking on a drag spot would result in three events
firing. In this order: mousedown, mouseup and click.
The code now recognizes if the event was meant
for expanding the column size by dragging or for sorting.