Fixed sort triggering when it shouldnt...
Beacuse of the way that all the header events interacted
with each other sort would trigger after dragging a column
and releasing over a column positioned before it.
We moved the handling of the sort event into mousedown.
This prevents weird bubbling issues when the click
event triggers after all the others.