Finished collapsible tree nodes...
There is a limitation though, when you collapse a node
and then collapse its parent, when you expand the
parent the child will lose its collapse status.
However since this isnt very important to me I will leave it as is for now.
Had to make changes to some function which involved traversing the tree.
now only nodes that are visible will be counted
towards length or get() etc.
This is necessary to access the correct node on row selected,
or to draw the correct length of tree lines.