Improved sort icon...
The icon is now black which stands out better and the
color is not a problem for people with color blindness.
The icon has been enlarged as well, since previously it
was rather small which made it hard to see especially
with its previous color.  Finally the code which decides
if the header has enough space for the sort icon now
takes into account if the cell is centerd or not.
It still doesnt care about right alignment but since its
not used anywhere in this application it is right now out of scope.