Fixed flickering images when redrawing...
While not a deal breaker this was a rather annoying visual.
My original attempt was to use double buffering.
Essentially first draw the contents to an offscreen canvas
and once all images have been loaded then draw that
buffer canvas onto the actual data layer.
This did not work, and turns out most browsers actually use double buffering.
I did not notice a difference after implementing this.
So the second and current implementation simply
reuses the image object for images that repeat -> diy caching.
The difference is quite staggering, not only does the
redraw happen faster the flickering is gone (mostly)
all this makes for a much smoother experience.
The flickering is not entirely gone actually.
It still happens when loading images for the first time.
This can be observed when switching to a weapon type
that has yet to be loaded, however after that first load
there will be no more flickering :)