Documentation and small restructure...
Moved clipRect in to CellRectangle class, this just makes
more sense since the method directly operates on rects
and further it help with moving some type info to the
new interface.

Change CellRenderer type info in the TreeView to use
the new CellRendererInterface.
This makes it more obvious that the base CellRenderer class
is just a base to built upon and not to be used directly,
and to lock it down with protected.
Further you can now easily determine without reading
docs on what your custom CellRenderer needs to implement.