Changed image to icon + important implementation details...
The semantic difference here is significant because
when we draw an icon we will assume that it has a 1:1
aspect ration and that it will always fit into the TreeView row.
So each icon will be scaled to 24px by 24px assuming default row_height.
However we don't know what the row height is unless
we have access to the treeview, and that is necessary
for calculating width. So for a uniform API we now
pass the TreeView in the CellRenderer constructor
and we can omit it later when calling draw().