A simple proof of concept for virtualization...
This draws only the rows of the TreeView that are
currently visible on the screen. But until this is usable
we need to draw the tree lines that come from the top
but who's origin point is not visible.
After that we can test out clipping again and optimize the drawing,
since right now we draw more than stricly necessary.
We also dont need to clear the entire canvas, just
the last part that was visible.