Add function for drawing lines, rename old line func...
Added an extra function to draw treelines to the parent,
and consequently renamed the old tree lines function
to drawTreeLinesToChildren to make it more obvious
what each function does.
Now this gets me to 90% of the solution, however thats
not quite good enough.
Right now if neither the child nor the parent are visible then the line wont be draw.
Which is not ideal. I think to really solve it with the least
amount of drawing  I need to draw the lines to children
from the current visible root nodes.