Added new tree lines implementation for virtualization...
This works, essentailly it checks all the root nodes of
the nodes that are going to be drawn and descends
all the unique root nodes and draws the lines for each
node stemming from those root nodes.
This means that we are not drawing the whole tree
anymore however this is still more expensive than
just drawing the parent and child lines.
When drawing the lines the first column is clipped so
 that in the future we can actually resize the first col
 without the lines being drawn over other columns.