Finally re-added clipping to each cell rect...
Previously I had tried to add it but it was just too slow
to justify its inclusion.
Now that drawing is virtualized we can afford the performance
hit since we are only drawing a part of the tree at a
time. This eliminates the need for the extreme elipsis
that is currently implemented.
However is still is probably a good idea to write a version
of drawText that draws as much text as it can and changes the last three
characters to "..." to indicate that there is more text that
is hidden by the clipped column.