Removed elipsis and added consideration for clipping...
So now instead of elipsizing content we simply clip.
The slight problem was that when the alignment was
other than default left the clip would not matter and
content would still be outside of the column.
Now when the content is too wide for the column we
change the alignment to Left and this gives a predicatable
content width which gets clipped exactly as it is supposed to.