Fixed issue with treelines not drawing correctly...
This would happen after the model was sorted which
would throw the iter out of wack. Since after sort
the iter was not updated, as a result this return bogus
results when trying to paint lines.
To fix this I implemented an ascend function on the model.
Which allows us to get the root node of the current
branch and from there we find out its index which
is then assigend to current_root.